Run Systemd tests only on linux
